:memo:[Add]신고추가(진행중), 블랙리스트 자동 등록(완료), 블랙리스트 카운팅(진행중)

+++ src/main/java/com/team3/itability/blackllist/dao/BlacklistDAO.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BlacklistDAO extends JpaRepository<BlacklistDTO, Integer> {
+public interface BlacklistDAO extends JpaRepository<BlacklistDTO, Long> {
 }
